+++ TransformData/transformResponseData.ts
@@ -1,12 +1,15 @@
 import { WhatsNewItem, WhatsNewSection } from "../data/whatsNewConfig";
-import { DashboardMetrics } from "../components/HomePage/ImpactGlance";
-import { CauseListItem, RecentCauseListResponse } from "../components/HomePage/NoticeAndCauseListSection";
+import {
+  CauseListItem,
+  RecentCauseListResponse,
+} from "../components/HomePage/NoticeAndCauseListSection";
+import { DashboardMetricsNew } from "../components/HomePage/Highlights";
 
 export interface MdmsApiResponse {
   MdmsRes: {
     LandingPage: {
       LatestAndComingSoon?: WhatsNewItem[];
-      DashboardMeterics?: DashboardMetrics[];
+      DashboardMeterics?: DashboardMetricsNew[];
     };
   };
 }
@@ -40,7 +43,7 @@ export function transformWhatsNewResponse(data: MdmsApiResponse): {
 }
 
 export function transformImpactGlance(data: MdmsApiResponse): {
-  stats: DashboardMetrics;
+  stats: DashboardMetricsNew;
 } {
   const allItems = data?.MdmsRes?.LandingPage?.DashboardMeterics || [];
   return {
@@ -53,21 +56,23 @@ export function transformCauseList(
 ): CauseListItem[] {
   const list = data?.RecentCauseList || [];
 
-  return list.map((item, index) => {
-    const date = new Date(item.date);
-    const options: Intl.DateTimeFormatOptions = {
-      day: "numeric",
-      month: "long",
-      year: "numeric",
-    };
-    const formattedDate = date.toLocaleDateString("en-IN", options);
-    const title = `${formattedDate} Causelist - 24x7 ON Court`;
+  return list
+    .map((item, index) => {
+      const date = new Date(item.date);
+      const options: Intl.DateTimeFormatOptions = {
+        day: "numeric",
+        month: "long",
+        year: "numeric",
+      };
+      const formattedDate = date.toLocaleDateString("en-IN", options);
+      const title = `${formattedDate} Causelist - 24x7 ON Court`;
 
-    return {
-      id: `${index + 1}`,
-      title,
-      date,
-      fileStoreId: item.fileStoreId,
-    };
-  }).sort((a, b) => b.date.getTime() - a.date.getTime());
+      return {
+        id: `${index + 1}`,
+        title,
+        date,
+        fileStoreId: item.fileStoreId,
+      };
+    })
+    .sort((a, b) => b.date.getTime() - a.date.getTime());
 }
