+++ pages/landing/index.tsx
@@ -1,9 +1,4 @@
 import React from "react";
-import NoticeAndCauseListSection from "../../components/HomePage/NoticeAndCauseListSection";
-import Support from "../../components/HomePage/Support";
-import ImpactGlance from "../../components/HomePage/ImpactGlance";
-import TestimonialCarousel from "../../components/HomePage/TestimonialCarousel";
-import WhatsNewSections from "../../components/HomePage/WhatNewSections";
 import StickyHelpButton from "../../components/Utils/StickyHelpButton";
 import HomeIntroVideo from "../../components/HomePage/HomeIntroVideo";
 import InfoBanner from "../../components/InfoBanner";
@@ -18,11 +13,6 @@ const Hero = () => {
       <HomeIntroVideo />
       <Highlights />
       <QuestionsSection />
-      {/* <NoticeAndCauseListSection /> */}
-      {/* <WhatsNewSections /> */}
-      {/* <ImpactGlance />
-      <TestimonialCarousel /> */}
-      {/* <Support /> */}
     </div>
   );
 };
